requirements.txt
 fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
sqlmodel>=0.0.14
psycopg2-binary>=2.9.7
pydantic>=2.4.0
pydantic[email]>=2.4.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pillow>=9.2.0
python-decouple>=3.8
pydantic-settings>=2.0.0

# AI/ML dependencies
sentence-transformers>=2.2.2
transformers>=4.21.0
torch>=1.12.0
numpy>=1.21.0
faiss-cpu>=1.7.4
scikit-learn>=1.1.0
opencv-python>=4.6.0

